Removed LanguageModelTokenizer. Logic moved into HFTransformersNLP, LMTokenizer tests adjusted and moved into HFTransformersNLP tests
