+++ rasa/nlu/utils/hugging_face/hf_transformers.py
@@ -1,14 +1,13 @@
 import logging
-from typing import Any, Dict, List, Text, Tuple, Optional
+from typing import Any, Dict, List, Text, Tuple, Optional, Type
 
 from rasa.core.utils import get_dict_hash
 from rasa.nlu.model import Metadata
-from rasa.nlu.tokenizers.whitespace_tokenizer import WhitespaceTokenizer
 from rasa.nlu.components import Component
 from rasa.nlu.config import RasaNLUModelConfig
 from rasa.shared.nlu.training_data.training_data import TrainingData
 from rasa.shared.nlu.training_data.message import Message
-from rasa.nlu.tokenizers.tokenizer import Token
+from rasa.nlu.tokenizers.tokenizer import Token, Tokenizer
 import rasa.utils.train_utils as train_utils
 import numpy as np
 
@@ -17,6 +16,7 @@ from rasa.nlu.constants import (
     DENSE_FEATURIZABLE_ATTRIBUTES,
     TOKEN_IDS,
     TOKENS,
+    TOKENS_NAMES,
     SENTENCE_FEATURES,
     SEQUENCE_FEATURES,
     NUMBER_OF_SUB_TOKENS,
@@ -55,6 +55,10 @@ class HFTransformersNLP(Component):
         "cache_dir": None,
     }
 
+    @classmethod
+    def required_components(cls) -> List[Type[Component]]:
+        return [Tokenizer]
+
     def __init__(
         self,
         component_config: Optional[Dict[Text, Any]] = None,
@@ -64,7 +68,6 @@ class HFTransformersNLP(Component):
 
         self._load_model_metadata()
         self._load_model_instance(skip_model_load)
-        self.whitespace_tokenizer = WhitespaceTokenizer()
 
     def _load_model_metadata(self) -> None:
 
@@ -241,7 +244,7 @@ class HFTransformersNLP(Component):
 
         Many language models add a special char in front of (some) words and split
         words into sub-words. To ensure the entity start and end values matches the
-        token values, tokenize the text first using the whitespace tokenizer. If
+        token values, use the tokens produced by the Tokenizer component. If
         individual tokens are split up into multiple tokens, we add this information
         to the respected token.
 
@@ -255,8 +258,7 @@ class HFTransformersNLP(Component):
             message.
         """
 
-        tokens_in = self.whitespace_tokenizer.tokenize(message, attribute)
-
+        tokens_in = message.get(TOKENS_NAMES[attribute])
         tokens_out = []
 
         token_ids_out = []
